Fixed some more warnings about data type sizes.
[TESTS]
Compile in VS 2013.

+++ backend/wbpublic/grt/common.cpp
@@ -285,7 +285,7 @@ namespace bec {
    * This position must be local to the enclosing group.
    */
   template <class T>
-  void move_list_ref_item(grt::ListRef<T> items, const grt::ValueRef &object, int to)
+  void move_list_ref_item(grt::ListRef<T> items, const grt::ValueRef &object, size_t to)
   {
     grt::Type object_type = object.type();
 
@@ -485,7 +485,7 @@ namespace bec {
 
   // Template instantiation to avoid having all this code in the header file.
   template WBPUBLICBACKEND_PUBLIC_FUNC void move_list_ref_item<db_mgmt_Connection>(grt::ListRef<db_mgmt_Connection> items,
-                                                       const grt::ValueRef &object, int to);
+                                                       const grt::ValueRef &object, size_t to);
   
   //------------------------------------------------------------------------------------------------
